diseño menú historical fragment
