+++ .idea/misc.xml
@@ -3,22 +3,35 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/bg_degrade_lightorange.xml" value="0.19722222222222222" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/bg_degrade_lightrasberry.xml" value="0.2169230769230769" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_historical_black_24dp.xml" value="0.27314814814814814" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_jogging_black_72dp.xml" value="0.19722222222222222" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.15885416666666666" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_profile_black_24dp.xml" value="0.27314814814814814" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_skiping_rope_72dp.xml" value="0.19722222222222222" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_skiping_rope_black_72dp.xml" value="0.19722222222222222" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_walking_72dp.xml" value="0.19722222222222222" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_walking_black_72dp.xml" value="0.19722222222222222" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_weight_24dp.xml" value="0.2169230769230769" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable-v24/ic_weight_72dp.xml" value="0.19722222222222222" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic__calendar_black_24dp.xml" value="0.2169230769230769" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_cake_black_24dp.xml" value="0.2169230769230769" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_calendar_black_24dp.xml" value="0.12314814814814815" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_climb_black_24dp.xml" value="0.12314814814814815" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_gender_black_24dp.xml" value="0.2169230769230769" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_height_black_24dp.xml" value="0.19722222222222222" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_invictus_color_100dp.xml" value="0.19722222222222222" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_launcher_background.xml" value="0.15885416666666666" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_run_black_24dp.xml" value="0.12314814814814815" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/drawable/ic_settings_black_24dp.xml" value="0.27314814814814814" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout-v31/fragment_profile.xml" value="0.2838541666666667" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/activity_main.xml" value="0.11775362318840579" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_dashboard.xml" value="0.14270833333333333" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_historical.xml" value="0.22" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_home.xml" value="0.1" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_notifications.xml" value="0.11775362318840579" />
-        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_profile.xml" value="0.25" />
+        <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_profile.xml" value="0.1" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/fragment_settings.xml" value="0.25" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/item_spinner.xml" value="0.14764492753623187" />
         <entry key="..\:/Users/Admin/Desktop/InvictusApp/Invictus/app/src/main/res/layout/root_preferences.xml" value="0.13722826086956522" />
