+++ app/src/main/res/drawable-v24/bg_degrade_lightorange.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient
+        android:startColor="#00FFFFFF"
+        android:endColor="@color/lightOrange"
+        android:angle="270" />
+</shape>
