+++ app/src/main/res/drawable-v24/ic_jogging_black_24dp.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512"
+    android:tint="?attr/colorControlNormal">
+    <path android:fillColor="@android:color/white"
+        android:pathData="M360.8,1.5c-14,3.5 -28.5,15 -34.8,27.4 -8.2,16.3 -7.6,36.1 1.5,51.7 6.2,10.6 20.1,21 32.5,24.4 11.1,3.1 26.9,1.1 38.5,-4.7 6.4,-3.3 18.5,-15.4 21.8,-21.8 5.8,-11.6 7.8,-27.4 4.7,-38.5 -3.4,-12.4 -13.8,-26.3 -24.4,-32.5 -11,-6.5 -28,-9 -39.8,-6z" android:strokeColor="#00000000"/>
+    <path android:fillColor="@android:color/white"
+        android:pathData="M207.9,64.9c-1.4,0.5 -3.6,1.4 -4.9,2.1 -4.8,2.6 -89.7,66.8 -91.8,69.5 -7.7,9.8 -4.3,25.4 6.9,31.2 4.9,2.7 13.1,3.1 18.3,0.9 1.8,-0.8 20.5,-14.2 41.4,-30l38.1,-28.5 27.3,13c15,7.2 27.7,13.4 28.1,13.8 0.4,0.4 -14.3,24.8 -32.6,54.2 -18.4,29.4 -34.4,55.3 -35.5,57.4 -6.6,13.1 -4.5,31.2 4.8,41.8 2.5,2.8 23.3,17.8 55.5,40.1 28.3,19.7 51.5,36.1 51.5,36.7 0,0.5 -13,26.9 -28.9,58.8 -16,31.8 -29.3,59.5 -29.6,61.5 -1,5.2 0.1,9.6 3.6,15 4.2,6.3 9.5,9 17.4,9 6.8,-0.1 10.5,-1.7 15.2,-6.6 1.9,-2 15.7,-28.4 36.2,-69.6 32.5,-65 33.1,-66.4 33.1,-72.3 0,-3.5 -0.7,-7.3 -1.6,-9.2 -0.9,-1.8 -17.1,-18.8 -36.1,-37.8l-34.6,-34.7 27.3,-46.1c15,-25.4 27.6,-46 28.1,-45.9 0.4,0.2 10.8,14 23.1,30.8 18.7,25.6 23,30.9 26.8,33l4.5,2.5 48.5,-0 48.5,-0 4.5,-2.5c5.2,-2.9 9.9,-9.9 10.7,-15.8 1.1,-8.4 -4.2,-17.7 -12.5,-21.8 -3.4,-1.7 -7.3,-1.9 -43.7,-2.4l-39.9,-0.5 -27.6,-37.6c-15.1,-20.6 -29.2,-39.5 -31.2,-42 -3.2,-3.8 -6.7,-5.9 -26.5,-15.6 -29.9,-14.7 -106.9,-50.8 -111.2,-52.2 -3.8,-1.2 -7.4,-1.3 -11.2,-0.2z" android:strokeColor="#00000000"/>
+    <path android:fillColor="@android:color/white"
+        android:pathData="M170.5,316l-12.5,24.9 -60.2,0.3c-56.6,0.3 -60.5,0.4 -63.8,2.2 -8.8,4.7 -12.4,10.9 -11.8,20.5 0.6,9.8 5.8,16.5 14.9,19.1 2,0.6 30.1,1 70.9,1 63.9,-0 67.7,-0.1 71.5,-1.9 2.5,-1.2 12.6,-10.6 27.5,-25.5 12.9,-12.9 23.2,-23.7 23,-23.9 -0.3,-0.2 -6.8,-4.6 -14.5,-9.8 -15.6,-10.5 -23.9,-18.2 -28.6,-26.4 -1.7,-3 -3.3,-5.5 -3.5,-5.5 -0.2,-0 -6,11.2 -12.9,25z" android:strokeColor="#00000000"/>
+</vector>
