+++ app/src/main/res/drawable-v24/ic_skiping_rope_black_24dp.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="200"
+    android:viewportHeight="200"
+    android:tint="?attr/colorControlNormal">
+    <path android:fillColor="@android:color/white"
+        android:pathData="M92.9,2.3c-11.2,7.4 -7.5,23.7 5.8,25.3 7.2,0.9 14,-4.7 14.9,-12.2 1,-7.6 -5.9,-15.4 -13.6,-15.4 -2.2,-0 -5.1,0.9 -7.1,2.3z" android:strokeColor="#00000000"/>
+    <path android:fillColor="@android:color/white"
+        android:pathData="M80.5,34.5c-5.5,2 -7.2,4.4 -13.9,20.2l-5.8,13.8 -10.5,10.7c-9.6,9.8 -10.5,11 -9.9,13.5 1,3.9 3,5.6 6.6,5.6 2.6,-0 4.9,-1.7 14.2,-10.9 9.5,-9.4 11.6,-12.1 15.2,-19.4l4.1,-8.5 0.3,46.9c0.2,46.6 0.2,47 2.4,49.3 3,3.2 8.1,3.1 11.3,-0.2l2.5,-2.4 0,-23.6 0,-23.5 3,-0 3,-0 0,36 0,36.1 2.5,2.4c3.2,3.3 8.3,3.4 11.3,0.2 2.2,-2.3 2.2,-2.3 2.4,-61.8l0.3,-59.4 4.1,8.5c3.6,7.3 5.7,10 15.2,19.4 9.3,9.2 11.6,10.9 14.2,10.9 3.6,-0 5.6,-1.7 6.6,-5.6 0.6,-2.5 -0.3,-3.7 -9.9,-13.5l-10.5,-10.7 -5.9,-13.9c-6.2,-15 -9.2,-18.9 -15.1,-20.6 -5.7,-1.5 -32.9,-1.2 -37.7,0.5z" android:strokeColor="#00000000"/>
+    <path android:fillColor="@android:color/white"
+        android:pathData="M44.5,114.7c1.7,27 4.9,42.2 11.5,55.9 10.4,21.6 29.1,31.9 51.6,28.7 10.7,-1.5 18.8,-5.8 26.3,-13.8 12.9,-13.9 20,-36.9 21.4,-69.8l0.4,-9.7 -2.7,-0 -2.7,-0 -0.7,12.7c-2.2,38.7 -12.5,62.2 -31.5,71.6 -6.2,3 -7.1,3.2 -18.1,3.2 -10.6,-0 -12,-0.2 -17.7,-2.9 -18.9,-8.7 -29.9,-33.5 -31.9,-72.2l-0.7,-12.4 -2.9,-0 -2.8,-0 0.5,8.7z" android:strokeColor="#00000000"/>
+</vector>
