+++ app/src/main/res/drawable/ic_climb_black_24dp.xml
@@ -1,0 +1,11 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:tint="?attr/colorControlNormal" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000000"
+        android:pathData="M161.6,1.6c-2,1.9 -2,4.8 -0.1,6.7 0.8,0.8 4.5,3.4 8.4,5.8l6.9,4.2 -0.4,17.6c-0.2,9.7 -0.7,31 -1,47.4l-0.6,29.9 -15.1,16.6c-8.2,9.2 -17.1,18.9 -19.5,21.6l-4.6,4.9 -26.5,2.8c-30,3.1 -31.1,3.3 -36.6,6.4 -9.7,5.6 -14.9,14.7 -14.9,26 0.1,3.8 0.4,9.2 0.8,11.9l0.8,4.9 11,-10.9c6.1,-6 10.8,-10.5 10.6,-9.9 -0.3,0.5 -9.8,16.4 -21.1,35.2 -25.1,41.7 -24.7,40.9 -24.7,47.5 0,4.6 1.3,7.9 11.6,29.3 24.7,50.8 30.2,61.2 33.6,63.8 9.1,6.7 20.3,6 28.2,-1.9 3.9,-3.9 5.6,-8.3 5.6,-14.6 0,-4.7 -1.5,-8.4 -16.5,-39 -9.1,-18.5 -16.6,-34.1 -16.7,-34.7 -0.3,-0.8 38.4,-68.1 38.9,-67.7 0.1,0.1 -3.8,14.8 -8.8,32.6l-9.1,32.3 10.3,21.5c5.7,11.8 11,21.7 11.7,22 0.8,0.3 13.9,-0.6 29.3,-1.8 20.2,-1.7 27.9,-2.7 27.9,-3.6 0,-1.8 -11.8,-117.3 -12.4,-121.5 -0.6,-3.4 0,-4.2 22,-28.5 12.4,-13.7 23.3,-26.3 24.2,-27.9 1.4,-2.5 1.8,-7.5 2.3,-30.5 0.4,-15.1 0.8,-33.6 0.8,-41l0.2,-13.5 3.8,4.3c6.1,7 8.7,13.5 19.5,48.6 7.1,23.2 11.2,35 13.5,38.9 2,3.2 16.8,19.3 36.9,39.8 18.4,18.9 35,36.7 36.9,39.5 5.3,8 7.5,14.8 8,24.9 0.2,4.9 -0.7,30.8 -2.1,57.5 -2.7,52.4 -2.6,55.4 2.2,66.2 2.5,5.7 42.7,77.1 46.6,82.8 7.3,10.8 11,13.2 46.6,30.8 37.6,18.6 39.1,19.1 44.1,13.4l2.9,-3.2 0,-240.5c0,-177.6 -0.3,-241.1 -1.1,-242.9 -2.7,-5.9 6.4,-5.6 -160.2,-5.6 -137.4,-0 -152.7,0.2 -154.1,1.6z" android:strokeColor="#00000000"/>
+    <path android:fillColor="#000000"
+        android:pathData="M98.5,59.6c-6.9,1.5 -11.4,2.9 -14.5,4.7 -21.5,12.5 -29,37.1 -17.8,59.1 2.7,5.2 10.3,13.2 15.6,16.3 26.9,15.8 60.5,-0.1 65.3,-30.9 3.5,-22.5 -12.2,-44.5 -35,-48.8 -5.4,-1 -9.9,-1.1 -13.6,-0.4z" android:strokeColor="#00000000"/>
+    <path android:fillColor="#000000"
+        android:pathData="M129.6,327.7c6.1,14.2 6.2,26.1 0.5,37.5 -1.6,3.2 -2.5,5.8 -2,5.8 0.5,-0 19.8,2.5 42.9,5.5l41.9,5.5 36.7,25.6c20.1,14.1 38.5,26.5 40.8,27.5 11.8,5.3 26.6,-0.4 32.2,-12.5 4.4,-9.5 2.2,-22 -5.2,-28.7 -6,-5.4 -83,-58.3 -86.8,-59.6 -2.9,-1.1 -99,-14.3 -103.5,-14.3 -0.5,-0 0.6,3.5 2.5,7.7z" android:strokeColor="#00000000"/>
+    <path android:fillColor="#000000"
+        android:pathData="M188,400c0,2.6 26.3,94.9 28.1,98.7 3.4,7 13.6,13.3 21.8,13.3 13.4,-0 25.1,-11.5 25.1,-24.8 0,-6.1 -13.4,-55.2 -15.3,-56.4 -0.7,-0.4 -10.5,-7.2 -21.9,-15.1 -19.4,-13.6 -21.1,-14.5 -26.5,-15.1 -3.2,-0.4 -7,-0.9 -8.5,-1.2 -1.7,-0.4 -2.8,-0.1 -2.8,0.6z" android:strokeColor="#00000000"/>
+</vector>
