+++ app/src/main/res/drawable/ic_run_black_24dp.xml
@@ -1,0 +1,11 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:tint="?attr/colorControlNormal" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000000"
+        android:pathData="M339,1.7c-26,4.3 -46,28.2 -46,54.6 0,14.6 5.5,27.5 16.4,38.3 22.4,22.4 57.8,21.6 79,-1.8 19.9,-21.9 19.1,-54.4 -1.9,-75.3 -3.8,-3.9 -9.3,-8.2 -12.2,-9.8 -9.7,-5.2 -24.9,-7.8 -35.3,-6z" android:strokeColor="#00000000"/>
+    <path android:fillColor="#000000"
+        android:pathData="M472.4,65.8c-6.5,2.3 -12.8,7.3 -15.8,12.5 -1.3,2.3 -8.1,19.9 -15.1,39.2 -6.9,19.2 -13,35.3 -13.4,35.8 -0.4,0.4 -11.9,-3.3 -25.6,-8.3l-24.9,-9.1 -8.1,2.3c-6.7,1.9 -10.3,2.2 -22,2.2 -12.2,-0.1 -15.1,-0.4 -22.8,-2.8 -19.6,-5.9 -35.8,-17.5 -46.5,-33.4 -4.1,-5.9 -4.2,-6 -14.4,-9.6 -22,-7.8 -33.9,-8.2 -80.3,-3 -35,3.9 -38.7,4.8 -48.2,11.8 -9.7,7.1 -12,11.5 -26.4,50.7 -14.4,39.1 -15.4,43.3 -12.5,53.4 2.3,7.5 10.6,16.3 18,18.8 5.5,1.9 13.3,2.2 18.6,0.7 5.2,-1.4 12.3,-6.2 15.5,-10.4 1.7,-2.2 7,-15.1 13.2,-31.9 5.7,-15.4 10.7,-28.4 11.1,-28.8 0.7,-0.6 59.2,-6.3 59.6,-5.7 0.2,0.2 -16,45.1 -35.9,99.8l-36.2,99.5 -42.4,0.5 -42.4,0.5 -5.6,2.8c-26.7,13 -26.5,50.9 0.3,63.3l6.3,2.9 52.4,0.3c35.3,0.2 54.2,-0.1 57.7,-0.8 11.7,-2.5 23.7,-11 29.6,-21.1 1.4,-2.4 6.5,-15.3 11.3,-28.6l8.8,-24.2 6.1,2.3c3.4,1.3 15.9,5.8 27.9,10.2 11.9,4.3 21.7,8.3 21.7,8.9 0,0.6 -8.2,23.3 -18.2,50.5l-18.2,49.5 0,9.5c-0.1,8.5 0.2,10.1 2.7,15 10,19.5 33.7,25.8 51.3,13.7 9.1,-6.3 9.9,-8 35,-76.9 12.9,-35.3 23.9,-67 24.5,-70.5 2.8,-16.1 -5.7,-35.1 -19.8,-44.1 -2.7,-1.7 -10.6,-5.2 -17.8,-7.8 -7.1,-2.6 -13.4,-5 -13.8,-5.4 -0.7,-0.6 36.6,-105.4 38.2,-107.2 0.3,-0.4 16,4.9 34.8,11.7 18.8,6.8 37,12.9 40.5,13.5 18,2.8 37.3,-7.1 45.6,-23.7 1.2,-2.3 9.5,-24.1 18.4,-48.5 15.5,-42.3 16.2,-44.6 16.2,-51.8 0.1,-9.1 -2.2,-14.8 -8.4,-21 -8,-8 -20.3,-10.9 -30.6,-7.2z" android:strokeColor="#00000000"/>
+    <path android:fillColor="#000000"
+        android:pathData="M7.8,223.1c-4.3,2.2 -7.8,8.1 -7.8,12.9 0,4.8 3.3,10.3 7.8,12.8 3.7,2.1 5.1,2.2 23.1,2.2 22.3,-0 25.5,-0.9 29.1,-8 2.5,-5 2.5,-9 0,-14 -3.6,-7.1 -6.8,-8 -29.2,-8 -17.2,0.1 -19.7,0.3 -23,2.1z" android:strokeColor="#00000000"/>
+    <path android:fillColor="#000000"
+        android:pathData="M7.8,283.1c-4.3,2.2 -7.8,8.1 -7.8,12.9 0,4.8 3.3,10.3 7.8,12.8 3.9,2.2 4.2,2.2 48.1,2.2 42.3,-0 44.3,-0.1 48.1,-2 4.5,-2.3 8,-8 8,-13 0,-5 -3.5,-10.7 -8,-13 -3.8,-1.9 -5.7,-2 -48.2,-2 -42.1,0.1 -44.5,0.2 -48,2.1z" android:strokeColor="#00000000"/>
+</vector>
